ArrayRadiation.jl Documentation

ArrayRadiation.DspUtility._mean — Method

_mean(x)::Real

Calculate the mean of a vector.

julia> using ArrayRadiation

julia> a = [1,2,3,4];

julia> ArrayRadiation.DspUtility._mean(a)
2.5

ArrayRadiation.DspUtility.antenna_matrix — Method

antenna_matrix(M, N, separation)

Generate locations (2D) for an evenly spaced MxN matrix of antenna elements, ensuring symmetry around the origin.

All elements are placed in the XY-plane.

Arguments

M : Number of elements along the Y-axis.

N : Number of elements along the X-axis.

separation : Distance between adjacent elements.

Example

```jldoctest julia> using ArrayRadiation

julia> DspUtility.antenna_matrix(1, 2, 0.5) 1×2 Matrix{Tuple{Float64, Float64, Float64}}: (-0.25, 0.0, 0.0) (0.25, 0.0, 0.0)

julia> DspUtility.antenna_matrix(1, 3, 1/2) 1×3 Matrix{Tuple{Float64, Float64, Float64}}: (-0.5, 0.0, 0.0) (0.0, 0.0, 0.0) (0.5, 0.0, 0.0)

julia> DspUtility.antenna_matrix(2, 1, 1/2) 2×1 Matrix{Tuple{Float64, Float64, Float64}}: (0.0, -0.25, 0.0) (0.0, 0.25, 0.0) ````

ArrayRadiation.DspUtility.db2mag — Method

Convert between decibal scale (magnitude) and linear

ArrayRadiation.DspUtility.db2pow — Method

Convert between decibal scale (power) and linear

ArrayRadiation.DspUtility.discard_low_values — Method

discard_low_values(scalar_value::Real, lower_limit::Real)

This function takes a scalar value and compares it against a specified lower limit. If the scalar value is below the lower_limit, the function returns nothing. Otherwise, it returns the scalar value unchanged.

Arguments

scalar_value::Real: The scalar value to compare.

lower_limit::Real: The threshold value below which the scalar is discarded.

Example

julia> using ArrayRadiation

julia> ArrayRadiation.DspUtility.discard_low_values(50, 40)
50

julia> ArrayRadiation.DspUtility.discard_low_values(30, 40)

ArrayRadiation.DspUtility.energy — Method

Return the signal energy in Joule

ArrayRadiation.DspUtility.linear_array — Method

Generate locations (1D) for a evenly spaced array of N elements. N is even.

Arguments

N Number of elements (assumed to be even).

separation Distance between adjacent elements.

ArrayRadiation.DspUtility.mag2db — Method

Convert between linear scale (magnitude) and decibel

ArrayRadiation.DspUtility.pow2db — Method

Convert between linear scale (magnitude) and decibel

ArrayRadiation.DspUtility.power_dBW — Method

Return the average signal power in dBW

ArrayRadiation.DspUtility.power_dBm — Method

Return the average signal power in dBm

ArrayRadiation.FrequencyDomain.x_dB_bandwidth — Method

Calculate the x dB bandwidth in number of bins.

julia> FrequencyDomain.x_dB_bandwidth([1,2,3,4,5,4,3,2,1], 3)
7

Arguments

spectrum_dB The signal in frequency domain.

x_dB The level of reduction in the band, e.g. half powe BW (3 dB).

ArrayRadiation.Kspace.azimuth2k_hat — Method

azimuth2k_hat(ϕ::Real)::Vector{Real}

Convert an azimuth angle ϕ (in radians) to a unit-length k-space vector in the XY-plane.
The resulting vector has the form [cos(ϕ), sin(ϕ), 0].

ArrayRadiation.Kspace.cos_taper — Function

Calculate cosine taper for elevation angle θ [rad]. Optionally provide α. α=1.4 accounts for mutual coupling between elements.

R. A. Dana, Electronically Scanned Arrays and K-Space Gain Formulation, Springer, 2019.

ArrayRadiation.Kspace.cos_taper_k_hat — Function

Calculate cosine taper for normalized k-space vector [kx, ky, kz]/|k|. Where |k| = 2π/λ0 Optionally provide α. α=1.4 accounts for mutual coupling between elements.

R. A. Dana, Electronically Scanned Arrays and K-Space Gain Formulation, Springer, 2019.

ArrayRadiation.Kspace.elevation2k_hat — Method

elevation2k_hat(θ::Real)::Vector{Real}

Convert an elevation angle θ (in radians) to a unit-length k-space vector in the XZ-plane.
The resulting vector has the form [sin(θ), 0, cos(θ)].

ArrayRadiation.Kspace.gain — Method

gain(k_xyz::AbstractVector{<:Real}, Ge::Real, r_xyz::AbstractArray, element_weights::AbstractArray)

Calculate GΩ(k), the angular domain gain in direction k for an array with specified element weights.

Arguments

k_xyz k-space vector.

Ge The antena gain in direction k_xyz [dB].

r_xyz The placement of each antenna element. in

element_weights The complex weight of each element. Must have the same size/shape as r_xyz.

References

R. A. Dana, Electronically Scanned Arrays and K-Space Gain Formulation, Springer, 2019.

ArrayRadiation.Kspace.gain_1D — Method

gain_1D(k_xyz::AbstractVector{<:Real}, Ge::Real, r_x::AbstractVector, element_weights::AbstractVector)

Calculate GΩ(k), the angular domain gain in direction k for an array with specified element weights.

Arguments

k_xyz k-space vector.

Ge The antena gain in direction k_xyz [dB].

r_x The placement of each antenna element. All elements are placed along the x-axis.

element_weights The complex weight of each element.

References

R. A. Dana, Electronically Scanned Arrays and K-Space Gain Formulation, Springer, 2019.

ArrayRadiation.Kspace.k2azimuth — Method

k2azimuth(k_xyz::Vector)::Real

Calculate the azimuth angle ϕ from a k-space vector.
The azimuth angle is the angle between the projection of the vector on the XY-plane and the x-axis.

ArrayRadiation.Kspace.k2elevation — Method

k2elevation(k_xyz::Vector)::Real

Calculate the elevation angle θ from a k-space vector.
The elevation angle is the angle between the z unit vector and the the k vector.

ArrayRadiation.Kspace.k_xyz — Method

k_xyz(θ::Real, ϕ::Real, λ0::Real)::Vector

Calculate k-space vector, representing antenna elevation and coordinates.

ArrayRadiation.Kspace.maximum_element_separation — Method

Calculate the maximum array element separation for maximum look angle. Result is fraction of wavelenth.

ArrayRadiation.Window.cosine_q — Function

cosine_q(M::Integer, q, scale = true)::AbstractVector

Create a coseine^q weighting (window).

Arguments

M The length of the window.

q For q=0 yields a uniform weight,q=1 yelds a cosine weighting, and q=2 yields a Hanning weighting.

scale Max of window is 1, when not scaled. sum( cosine_q(M, q) ) = M when scaled.

Example

julia> using ArrayRadiation

julia> N = 64;

julia> W = Window.cosine_q(N, 1);

julia> round.( W[1:4], sigdigits=3 )
4-element Vector{Float64}:
 0.0385
 0.116
 0.192
 0.269

julia> round( sum(W), sigdigits=2)
64.0

References

S. Yan, Broadband Array Processing, Springer, 2019

ArrayRadiation.Window.rangeOmit — Method

Returns 1:n but omits m.

ArrayRadiation.Window.split_window — Method

split_window(W::AbstractVector)::AbstractVector

Split a window to create a difference beam

Arguments

W The window to split.

Example

julia> using ArrayRadiation

julia> N = 64;

julia> W = Window.taylor(N, 4, -35);

julia> W = Window.split_window(W);

julia> round.( W[1:4], sigdigits=3 )
4-element Vector{Float64}:
 0.28
 0.288
 0.305
 0.329

julia> round.( W[N-3:N], sigdigits=3 )
4-element Vector{Float64}:
 -0.329
 -0.305
 -0.288
 -0.28

Below is an example of a split Taylor window.

ArrayRadiation.Window.taylor — Method

taylor(N::Integer, n_bar::Integer, sll_dB::Real)::AbstractVector

Create taylor weighting (window).

Arguments

N The window length.

n_bar The number of nearly constant-level sidelobes adjacent to the main lobe.

sll_dB Peak sidelobe_level in dB.

Example

julia> using ArrayRadiation

julia> W = Window.taylor(64, 4, -35);

julia> round.( W[1:4], sigdigits=3 )
4-element Vector{Float64}:
 0.28
 0.288
 0.305
 0.329

julia> round( sum(W), sigdigits=5)
64.0

References

Carrar, Goodman and Majewski, Spotlight Synthetic Aperture Radar: Signal Processing Algorithms, Artech House, 1995