ArrayRadiation.jl Documentation

ArrayRadiation.DspUtility._mean — Method

_mean(x)::Real

Calculate the mean of a vector.

julia> using ArrayRadiation

julia> a = [1,2,3,4];

julia> ArrayRadiation.DspUtility._mean(a)
2.5

ArrayRadiation.DspUtility.db2mag — Method

Convert between decibal scale (magnitude) and linear

ArrayRadiation.DspUtility.db2pow — Method

Convert between decibal scale (power) and linear

ArrayRadiation.DspUtility.energy — Method

Return the signal energy in Joule

ArrayRadiation.DspUtility.linear_array — Method

Generate locations (1D) for an x*λ spaced array of N elements. N is even.

ArrayRadiation.DspUtility.mag2db — Method

Convert between linear scale (magnitude) and decibel

ArrayRadiation.DspUtility.pow2db — Method

Convert between linear scale (magnitude) and decibel

ArrayRadiation.DspUtility.power_dBW — Method

Return the average signal power in dBW

ArrayRadiation.DspUtility.power_dBm — Method

Return the average signal power in dBm

ArrayRadiation.FrequencyDomain.x_dB_bandwidth — Method

Calculate the x dB bandwidth in number of bins.

julia> FrequencyDomain.x_dB_bandwidth([1,2,3,4,5,4,3,2,1], 3)
7

Arguments

spectrum_dB The signal in frequency domain.

x_dB The level of reduction in the band, e.g. half powe BW (3 dB).

ArrayRadiation.Kspace.cos_taper — Function

Calculate 1D cosine taper for angle θ [rad]. Optionally provide α. α=1.4 accounts for mutual coupling between elements.

R. A. Dana, Electronically Scanned Arrays and K-Space Gain Formulation, Springer, 2019.

ArrayRadiation.Kspace.gain_2D — Function

Calculate GΩ(k), the angular domain gain at angle k [rad] for an array with specified element weights.

R. A. Dana, Electronically Scanned Arrays and K-Space Gain Formulation, Springer, 2019.

Arguments

k Angle [rad].

element_gain The antena gain in direction k [dB].

element_placement_λ The placement of each antenna element in wavelength units.

element_weighting The complex weight of each element.

α Cosine taper/mutual coupling coefficient α>1.

ArrayRadiation.Kspace.maximum_element_separation — Method

Calculate the maximum array element separation for maximum look angle. Result is fraction of wavelenth.

ArrayRadiation.Window.cosine_q — Function

cosine_q(M::Integer, q, scale = true)::AbstractVector

Create a coseine^q weighting (window).

Arguments

M The length of the window.

q For q=0 yields a uniform weight,q=1 yelds a cosine weighting, and q=2 yields a Hanning weighting.

scale Max of window is 1, when not scaled. sum( cosine_q(M, q) ) = M when scaled.

Example

julia> using ArrayRadiation

julia> N = 64;

julia> W = Window.cosine_q(N, 1);

julia> round.( W[1:4], sigdigits=3 )
4-element Vector{Float64}:
 0.0385
 0.116
 0.192
 0.269

julia> round( sum(W), sigdigits=2)
64.0

References

S. Yan, Broadband Array Processing, Springer, 2019

ArrayRadiation.Window.rangeOmit — Method

Returns 1:n but omits m.

ArrayRadiation.Window.split_window — Method

split_window(W::AbstractVector)::AbstractVector

Split a window to create a difference beam

Arguments

W The window to split.

Example

julia> using ArrayRadiation

julia> N = 64;

julia> W = Window.taylor(N, 4, -35);

julia> W = Window.split_window(W);

julia> round.( W[1:4], sigdigits=3 )
4-element Vector{Float64}:
 0.28
 0.288
 0.305
 0.329

julia> round.( W[N-3:N], sigdigits=3 )
4-element Vector{Float64}:
 -0.329
 -0.305
 -0.288
 -0.28

Below is an example of a split Taylor window.

ArrayRadiation.Window.taylor — Method

taylor(N::Integer, n_bar::Integer, sll_dB::Real)::AbstractVector

Create taylor weighting (window).

Arguments

N The window length.

n_bar The number of nearly constant-level sidelobes adjacent to the main lobe.

sll_dB Peak sidelobe_level in dB.

Example

julia> using ArrayRadiation

julia> W = Window.taylor(64, 4, -35);

julia> round.( W[1:4], sigdigits=3 )
4-element Vector{Float64}:
 0.28
 0.288
 0.305
 0.329

julia> round( sum(W), sigdigits=5)
64.0

References

Carrar, Goodman and Majewski, Spotlight Synthetic Aperture Radar: Signal Processing Algorithms, Artech House, 1995